--- tests/pytest_tests/unit_tests_old/test_launch/test_wandb_agent.py
@@ -1,33 +1,0 @@
-"""Agent tests"""
-import wandb
-
-
-def test_agent_ignore(live_mock_server):
-    sweep_entities = []
-    sweep_projects = []
-
-    def train():
-        run = wandb.init(entity="ign", project="ignored")
-        sweep_projects.append(run.project)
-        sweep_entities.append(run.entity)
-        run.finish()
-
-    wandb.agent("test-sweep-id-3", function=train, count=1)
-
-    assert len(sweep_projects) == len(sweep_entities) == 1
-    assert sweep_projects[0] == "test"
-    assert sweep_entities[0] == "mock_server_entity"
-
-
-def test_agent_ignore_runid(live_mock_server):
-    sweep_run_ids = []
-
-    def train():
-        run = wandb.init(id="ignored")
-        sweep_run_ids.append(run.id)
-        run.finish()
-
-    wandb.agent("test-sweep-id-3", function=train, count=1)
-
-    assert len(sweep_run_ids) == 1
-    assert sweep_run_ids[0] == "mocker-sweep-run-x91"
